Move ratio calc to a custom hook & Do a bit of TypeScript magik
