+++ src/types/photoResizeResult.ts
@@ -1,0 +1,8 @@
+export type PhotoResizeResult = {
+  uri: string;
+  path: string;
+  name: string;
+  size: number;
+  width: number;
+  height: number;
+};
